.bashrc gpack: Drop --aggressive, but allow it to be passed in
We don't want always to do --aggressive because it can be one or two orders
of magntitude slower on large repos when there's relatively little to
compress since the last GC. So we drop the arg, but you can still pass it
in on the command line.

This is not really the way we want to do the args; you should be able to
pass in the `git` args for all commands (e.g., --git-dir), which you can't
right now, and we should extract --aggressive from that to separately pass
to `gc`.