bin/gitcmd-short.bash: `source`able file with command definitions
Moving the (function) command definitions to a file in bin/ means that you
can `source` it from anywhere without worrying about paths. (`source`
searches $PATH just like a stand-alone command would.) Note that this file
is _not_ executable because it doesn't work except when sourced. (This also
keeps it out of command-line completion, which doesn't show non-executable
files even when in a directory in $PATH.)

For users of dot-home, this works just as before; the dot/bashrc.inb4
fragment becomes simply `source gitcmd-short.bash`.

For people not using dot-home, they can manually link or copy the bin/
files to a directory in their $PATH (or add the repo's bin/ dir to their
$PATH) and then add just `source gitcmd-short.bash` to their ~/.bashrc.