glr/glra → lr/lrh: better names and docs for 'list refs' functionality
Some experience has shown that `glr` is hard to remember and type, and the
conflict with the `lr` ("ls recursive") program isn't as bad as it seems,
since it doesn't affect scripts. (Interactive shell users will have to
rename one or the other; we suggest `lrr() { command lr "$@"; }`)

We also change `a` for "all" to `h` for "all heads," since we want to
extend this to be able to list really all references, including non-"heads"
unknown to `git branch`, with `lra`.

Also, expand the documentation quite a bit, including the completely
forgotten "restrict to refs matching any of these strings" functionality.