gitcmd-short: Add -X (no switch to alt screen) to LESS options
Without -X, less will send a termincap init string that switches to the
alternate screen, meaning that when you exit and it switches back to the
regular screen you will no longer see the output. This is rather annoying
with e.g. `logb -5` when you want to see and cut/paste the commit IDs for
further commands.

So when we specify LESS, we now add -X. We don't need to do this where we
don't specify LESS because if `git` detects no LESS environment variable
(as is the case for most users), it appears to generate a LESS=-X or
similar automatically.

Reviewed-by: Curt J. Sampson <cjs@cynic.net>