bin/gitcmd-abbrev fetch(): Remove more cruft from output
We signifciantly reduce the amount of output, limiting it now to just
one line per remote plus one line per actually updated branch from
each remote.

This removes the multiple 'remote: ...' lines printed on stderr
through the simple expedient of sending stderr to a non-terminal,
which causes Git not to print those lines. This has the unfortunate
side-effect of changing all stderr output to go to stdout instead, but
is livable since `fetch` is mainly used interactively rather than in
scripts. But script writers must beware of this!

We also join 'Fetching <remote-name>' lines with their following
'From <remote-URL>' lines, when the latter are present. (The latter
will not appear if we're already fully up to date from that remote.)