bin/gitcmd-abbrev fetch(): Add multi-repo fetch
Now with any argument that's a path to a directory containing a .git/
subdirectory, assume all args are paths to repos to be fetched using
fetch() with no arguments. Any arguments that are not such a path are
ignored. The user is informed for each one whether it's being feched
or ignored.