bin/st: Fix error when submodules are present and run from subdir
When run from a subdir and submodules were found, `st` tried to `cd`
to them relative to the current path, which failed if the CWD was
not the root of the repo.

Though this is fixed, the paths it prints are still relative to the repo
root, whereas it might make more sense to print them relative to the
current directory.