bin/st: Set -Eeuo pipefail
The only thing failing under the new strict options was one call to
vcs_for_root where we don't care about the exit code.

This has been tested only with Git, but support for other VCSs is
currently mostly in an "XXX write me" state anyway. (That code is
probably rotting, and we should consider whether we even want to keep
it.)