bin/st: Remove support for Git <1.7.9 that has no `status -b` option
This is extra, confusing code that's not likely to be needed any more now
that we're many, many years past Git 1.x. Further, it was preventing us
from using `-o pipefail` because `git status -h` exits with code 129 that
we then would have to explicitly ignore, making the code messier yet. All
in all, too expensive to maintain for the potential benefit.