bin/st: Handle missing submodules
A submodule is "missing" in, e.g., a fresh clone done without
`--recursive` where someone hasn't run `git submodule update --init`.

Previously when a submodule was missing, we'd try to run `st` in its
subdir anyway, which would search upwards for .git, find the one in
the parent repo, run `st` on that, which would try to show status of
the submodule dirs, and continue to recurse infinitely.

We now check first if the subdir has a `.git` file or directory and,
do a recursive `st` in it only if it does, so that it will always work
properly. Otherwise we just print `MISSING` as the status information
for that submodule.