gitcmd-abbrev mbase() grmu(): Change to use `main`; fix bug in mbase()
Add back the `-` that mbase() somehow lost in the handling of its second
parameter causing it not to set the default value properly.

Both these now use `main` instead of `master`, because the former is more
correct and now more common. Still, would be nice if we figured out a way
to get the proper "main" branch we should be using. Perhaps look up the
default head of the remote that the tracking branch of the current branch
is using?