bin/st: Better submodule status recursion, now defaulting to 1 level
We now take -0, -1 (default), -2 etc. to determine the number of levels of
submodules to recurse, and change the length of the `__` prefix to indicate
submodule depth (which was previously incomprehensible).

This also changes the recursive call to `st` to "$0" so it will work
even if called directly when it's not in the path.