pfetch(): Renamed from fetchp() to improve command-line completion
Rather than typing `fe→p↵` you can now just type `pfe→↵`, a pattern
matching the common `fe→↵`.

Arguably `pushf` should be changed to `fpush` to match the above, but
it doesn't provide the same benefit because I don't use completion on
those commands. (The 4/5 character break seems to be the point at
which we switch to using completion, though it may also depend on the
fingering one needs to use for the keystrokes in the command.) But
also we do want to encourage the use of `push` over `pushf` wherever
the former can be used.