fetch(): Do not run `git status` if we're not in a work tree
Bare repos don't have a work tree, and so `git status` produces an
error. So don't try to run it if we're not in a work tree, as shown by
`git rev-parse --is-inside-work-tree`.

I'm not sure how new the `--is-inside-work-tree` command is, but until
we get a test framework set up and can actually test this with
different versions of Git, I'll assume it works anywhere people are
likely to be using it, given that I'm generally testing on Debian
which doesn't generally have a very recent release of Git. (2.39.5 in
this case.)