XXX comma-args: Shorter way of specifying some common long arguments
This doesn't actually work the way we need it to; you don't want to
use this commit unless you're trying to look at how it's broken.
Instead use the following commmit, which fixes this. This commit is
retained only to document the broken approach, since it may have ideas
that are usable in other circumstances.

This allows you to add e.g. `,dr` as your first argument to a log
command to add the --no-decorate and --reverse options. See the README
for details. This just applies to log-related commands right now.

logb() has been changed to remove --graph, since we often want to be
able to reverse it (',r'), and that can't be done with --graph.
logab() does add --graph, however, since we generally want to see that
in graph format and reversing it doesn't make much sense anyway.

Implementation Notes:

* During testing, it's feeling like it would be nice to be able to use
  comma-args anywhere in the command line, rather than just at the
  start, but that creates the problem of how we figure out which
  element of $@ to remove.

* Though it would be convenient, __gitcmdabbrev_cparse() cannot just
  echo "shift" or ":" so it could be called with
  `$(__gitcmdabbrev_cparse "$@")`, since that unfortunately starts a
  subshell and leaves the function unable to modify global variables.
  There doesn't really seem to be any way in Bash to call a function
  and easily capture the output without starting a subshell.

* Working without global variables, but instead passing back (in a
  `$(...)` result that is executed) doesn't work well because the
  extra layer of evaulation causes array elements with spaces in them
  to be split into two arguments. This actually might not really be a
  problem, but for the moment it seems better to err on the safe side.

* As noted in the comments, we parse the comma-args once, store them
  in a global, and then pass a plain ',' arg on to any other functions
  we call, which indicates that those functions should leave the
  global alone rather than clearing it and re-parsing. Only the last
  function that actually invokes `git` adds "${__gitcmdabbrev_cargs[@]}"
  to the command line. Another way of doing it would be to have each
  function use `... "${__gitcmdabbrev_cargs[@]}" "$@"` to add those
  to the $@ of called function, but that's not only more verbose, but
  also opens the possibility of someone doing `logb ,a ,b` and having
  several sets of comma-args parsed (maybe, depending on how many
  there are and how many nested function calls there are), which does
  not seem like a good idea.