bin/gitcmd-abbrev: Entirely rework comma-args to fix positioning
As it turned out, delaying the addition of comma-args until the final
call to `git` doesn't work, because functions will add things like
`--decorate=short` to the front of $@ and that will override the
earlier --no-decorate added to $__gitcmdabbrev_cargs by the ',d'
option. So we must place the comma-generated args between the default
prefixed args of things like logb() and the "$@" that it passes on to
other functions.

So we take the easy route out and simply return commands to be
eval'd that set a local $cargs[] array, as well as doing the
shift of $@ if necessary.

This currently prints the setting of cargs in a way that does not
allow arguments with spaces in them; these will end up being split
into two arguments. This shouldn't be an issue for what we're doing
here, but if it becomes one we can use a more sophisticated routine to
quote the arguments that it's printing. (There may also be other eval
gotchas here that we may need to fix at some point if they arise.)

The message for the previous, broken commit, contains further useful
information about implementation strategies and changes not related
to this fix, such as removing --graph from logb(). The previous
commit also contains the README updates for comma-args.