gitcmd-abbrev comma-opts: Add digits [0-9] for -0 through -9 args
The `-#` arguments are typically given to `log` for the number of
commits to show.

This supports only up to 9; ideally we should turn sequences of
consecutive digits into multi-digit arguments (i.e., ,99r should
give `-99 --reverse`), but this is not nearly as simple to do.

(One way to do this might be to have an extra variable outside the
loop that keeps track of the current sequence of digits, if any. If
there is a current sequence and we get another digit char, append the
char to the current sequence and replace the last argument in the
array with a new one with the new sequence of digits.)