`pushu` does better error checks and always uses current local branch
The most important change here is that `pushu` will now fail if the
branch you're trying to push already has an upstream set. Previously
this would unset the upstream if you already had one and leave you in
a confused state. If you really want to override this, you can do so
with the new `-f` option, which will unset the upstream and properly
reset it to the new one now.

It also removes the ability to push an arbitrary branch (i.e., not the
current branch) as this was little used; instead the branch being
pushed is always the current branch and the third argument is the name
of the branch on the remote, for those who use different (e.g.,
shorter) local branch names from the remote branch name.