bin/gctr: Allow any abbreviation of trailer line type names
We now accept the actual trailer name or any abbreviation down to the
first letter of it, in upper or lower case.

There are currently no collisions in the first letter so all trailers
can be selected with just a single letter. In the future we expect
that in the case of a collision the single letter will be assigned to
the more common one, and perhaps an alternate single letter can also
be used for the other one.