bin/{gctr → gt}: Shorten name of trailer line generator
This can be used quite often (even multiple times per commit) by those
who use trailer lines, so a name that's both easier to remember and
faster to type is a boon.

This conflicts with `/usr/bin/gt` from the Debian `genometools`
package, which is a pretty niche thing so we're willing to live with
the conflict.